feature #21: 서비스 함축
