+++ app/src/main/java/com/eoyeongbooyeong/booyoungee/di/RepositoryModule.kt
@@ -1,9 +1,7 @@
 package com.eoyeongbooyeong.booyoungee.di
 
 import com.eoyeongbooyeong.data.repositoryImpl.AuthRepositoryImpl
-import com.eoyeongbooyeong.data.repositoryImpl.TokenReissueRepositoryImpl
 import com.eoyeongbooyeong.domain.repository.AuthRepository
-import com.eoyeongbooyeong.domain.repository.TokenReissueRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -18,10 +16,4 @@ internal abstract class RepositoryModule {
     abstract fun provideAuthRepository(
         authRepositoryImpl: AuthRepositoryImpl,
     ): AuthRepository
-
-    @Binds
-    @Singleton
-    abstract fun provideTokenReissueRepository(
-        tokenRepositoryImpl: TokenReissueRepositoryImpl,
-    ): TokenReissueRepository
 }
