+++ app/src/main/java/com/eoyeongbooyeong/booyoungee/di/ServiceModule.kt
@@ -2,7 +2,6 @@ package com.eoyeongbooyeong.booyoungee.di
 
 import com.eoyeongbooyeong.booyoungee.di.qualifier.NoToken
 import com.eoyeongbooyeong.data.service.AuthService
-import com.eoyeongbooyeong.data.service.TokenReissueService
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -13,11 +12,6 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 object ServiceModule {
-    @Provides
-    @Singleton
-    fun provideTokenReissueService(@NoToken retrofit: Retrofit): TokenReissueService =
-        retrofit.create(TokenReissueService::class.java)
-
     @Provides
     @Singleton
     fun provideAuthService(@NoToken retrofit: Retrofit): AuthService =
