+++ data/src/main/java/com/eoyeongbooyeong/data/datasource/AuthDataSource.kt
@@ -1,5 +1,9 @@
 package com.eoyeongbooyeong.data.datasource
 
-interface AuthDataSource {
+import com.eoyeongbooyeong.domain.model.TokenModel
 
+interface AuthDataSource {
+    suspend fun reissueTokens(
+        refreshToken: String
+    ): Result<TokenModel>
 }
