+++ data/src/main/java/com/eoyeongbooyeong/data/service/AuthService.kt
@@ -1,4 +1,11 @@
 package com.eoyeongbooyeong.data.service
 
+import com.eoyeongbooyeong.domain.model.TokenModel
+import retrofit2.http.POST
+
 interface AuthService {
+    @POST("token/reissue")
+    suspend fun reissueTokens(
+        refreshToken: String,
+    ): TokenModel
 }
