+++ domain/src/main/java/com/eoyeongbooyeong/domain/repository/AuthRepository.kt
@@ -1,4 +1,9 @@
 package com.eoyeongbooyeong.domain.repository
 
+import com.eoyeongbooyeong.domain.model.TokenModel
+
 interface AuthRepository {
+    suspend fun reissueTokens(
+        refreshToken: String
+    ): Result<TokenModel>
 }
