--- domain/src/main/java/com/eoyeongbooyeong/domain/repository/TokenReissueRepository.kt
@@ -1,9 +1,0 @@
-package com.eoyeongbooyeong.domain.repository
-
-import com.eoyeongbooyeong.domain.model.TokenModel
-
-interface TokenReissueRepository {
-    suspend fun reissueTokens(
-        refreshToken: String
-    ): Result<TokenModel>
-}
